refactor: the inspector owns the compiler API, not the CLI
`@pikku/cli` imported `typescript` in three places, so installing the CLI
pulled in a second copy of the compiler even though the inspector — the
package whose whole job is reading TypeScript — already had one.

All three move behind the inspector:

  - `collectSurface` was self-contained (node builtins and `typescript`, and
    nothing else), so the file moves wholesale to a new `@pikku/inspector/surface`
    entry point. Its 23 tests move with it.
  - `pikku-barrel-checks` only needed "the module specifiers this file imports
    from", now `readModuleSpecifiers`. The rest of that check — findings,
    messages, fix hints — is CLI domain and stays.
  - `dist` only needed a tsconfig's resolved `outDir`, now `readTsconfigOutDir`.

`typescript` drops to a devDependency on the CLI, which still uses it in tests
to compile generated output. Nothing about the 6/7 split changes: the inspector
stays on 6 because 7 ships no compiler API.
