+++ packages/cli/package.json
@@ -63,7 +63,6 @@
     "tinyglobby": "^0.2.17",
     "ts-json-schema-generator": "2.9.1-next.18",
     "tsx": "^4.23.12",
-    "typescript": "^6.0.3",
     "yaml": "^2.9.0",
     "zod": "^4.4.3",
     "zod-to-ts": "^2.1.0"
@@ -86,6 +85,7 @@
     "@types/pg": "^8.21.0",
     "@types/semver": "^7.8.0",
     "@types/ws": "^8",
+    "typescript": "^6.0.3",
     "ws": "^8.21.3"
   },
   "files": [
