+++ packages/inspector/src/index.ts
@@ -64,3 +64,5 @@ export type {
   SerializedWorkflowGraphs,
 } from './utils/workflow/graph/index.js'
 export { resolveCoreType } from './utils/resolve-core-type.js'
+export { readModuleSpecifiers } from './utils/read-module-specifiers.js'
+export { readTsconfigOutDir } from './utils/read-tsconfig-out-dir.js'
