+++ packages/inspector/src/surface/collect-surface.test.ts
