+++ packages/inspector/src/surface/collect-surface.ts
